Rename validation folders to validator for consistency
- Rename validation-request to validator-request
- Rename validation-response to validator-response
- Rename validator.ts to request-validator.ts in validator-request
- Update all import paths across the codebase
- Maintain consistent naming pattern with schema-request/schema-response
