+++ packages/kori/src/kori/kori-internal.ts
@@ -7,8 +7,8 @@ import { type KoriPlugin } from '../plugin/index.js';
 import { type KoriRouter, type KoriRouterHandler } from '../router/index.js';
 import { type KoriRequestSchemaDefault } from '../schema-request/index.js';
 import { type KoriResponseSchemaDefault } from '../schema-response/index.js';
-import { type KoriRequestValidatorDefault } from '../validation-request/index.js';
-import { type KoriResponseValidatorDefault } from '../validation-response/index.js';
+import { type KoriRequestValidatorDefault } from '../validator-request/index.js';
+import { type KoriResponseValidatorDefault } from '../validator-response/index.js';
 
 import { createFetchHandler } from './fetch-handler-factory.js';
 import { type Kori, type KoriRouteDefinition } from './kori.js';
