+++ packages/kori/src/validator-request/index.ts
@@ -13,4 +13,4 @@ export {
   type KoriRequestValidator,
   type KoriRequestValidatorDefault,
   type KoriRequestValidatorMethods,
-} from './validator.js';
+} from './request-validator.js';
